// src/shared/api/models/getApiProductCategories200Item.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiProductCategories200ItemDescription } from './getApiProductCategories200ItemDescription';

export type GetApiProductCategories200Item = {
    id: string;
    /**
     * @minimum -2147483648
     * @maximum 2147483647
     */
    seqId: number;
    name: string;
    description: GetApiProductCategories200ItemDescription;
    createdAt: string;
    updatedAt: string;
};