// src/shared/api/models/getApiProductsListing200.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiProductsListing200ProductsItem } from './getApiProductsListing200ProductsItem';
import type { GetApiProductsListing200ServicesItem } from './getApiProductsListing200ServicesItem';

export type GetApiProductsListing200 = {
    products: GetApiProductsListing200ProductsItem[];
    services: GetApiProductsListing200ServicesItem[];
};