// src/shared/api/models/putApiProductCategoriesId200.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { PutApiProductCategoriesId200Description } from './putApiProductCategoriesId200Description';

export type PutApiProductCategoriesId200 = {
    id: string;
    /**
     * @minimum -2147483648
     * @maximum 2147483647
     */
    seqId: number;
    name: string;
    description: PutApiProductCategoriesId200Description;
    createdAt: string;
    updatedAt: string;
};